+++ parler/tests/testapp/models.py
@@ -86,3 +86,42 @@ class UniqueTogetherModel(TranslatableModel):
             ]
         }
     )
+
+
+class Level1(TranslatableModel):
+    l1_translations = TranslatedFields(
+        l1_title = models.CharField(max_length=200)
+    )
+
+
+class Level2(Level1):
+    l2_translations = TranslatedFields(
+        l2_title = models.CharField(max_length=200)
+    )
+
+
+class ProxyBase(TranslatableModel):
+    base_translations = TranslatedFields(
+        base_title = models.CharField(max_length=200)
+    )
+
+
+class ProxyModel(ProxyBase):
+    proxy_translations = TranslatedFields(
+        proxy_title = models.CharField(max_length=200)
+    )
+
+    class Meta:
+        proxy = True
+
+
+class DoubleModel(TranslatableModel):
+    shared = models.CharField(max_length=200, default='')
+
+class DoubleModelTranslations(TranslatedFieldsModel):
+    master = models.ForeignKey(DoubleModel, related_name='base_translations')
+    l1_title = models.CharField(max_length=200)
+
+class DoubleModelMoreTranslations(TranslatedFieldsModel):
+    master = models.ForeignKey(DoubleModel, related_name='more_translations')
+    l2_title = models.CharField(max_length=200)